third_party/WebKit/LayoutTests/http/tests/inspector-enabled/console-exception-source-url.html

<html>
<head>
<script src="../inspector/inspector-test.js"></script>
<script src="../inspector/console-test.js"></script>
<script type="text/javascript">
function onload()
{
    if (window.testRunner)
        testRunner.waitUntilDone();
    setTimeout(showInspectorAndRunTest, 0);

    var scriptContents = "function foo() { FAIL.FAIL; } \n //# sourceURL=foo2.js";
    eval(scriptContents);
    foo();
}


function showInspectorAndRunTest()
{
    if (window.testRunner)
        testRunner.showWebInspector();
    runTest();
}

function test()
{
    InspectorTest.expandConsoleMessages();
    InspectorTest.dumpConsoleMessages();
    InspectorTest.addResult("TEST COMPLETE");
    InspectorTest.completeTest();
}
</script>
</head>
<body onload="onload()">
<p>Tests that when exception happens before inspector is open source url is correctly shown in console.</p>
</body>
</html>
